+++ lib/cypress_dev.rb
@@ -1,4 +1,0 @@
-require 'cypress_dev/version'
-require 'cypress_dev/configuration'
-require_relative './cypress_dev/railtie' if defined?(Rails)
-