+++ lib/cypress_on_rails.rb
@@ -1,0 +1,4 @@
+require 'cypress_on_rails/version'
+require 'cypress_on_rails/configuration'
+require_relative './cypress_on_rails/railtie' if defined?(Rails)
+