+++ lib/generators/cypress_on_rails/templates/spec/cypress/cypress_helper.rb
@@ -17,13 +17,13 @@ rescue LoadError => e
   end
 end
 
-require 'cypress_dev/smart_factory_wrapper'
+require 'cypress_on_rails/smart_factory_wrapper'
 
-factory = CypressDev::SimpleRailsFactory
+factory = CypressOnRails::SimpleRailsFactory
 factory = FactoryBot if defined?(FactoryBot)
 factory = FactoryGirl if defined?(FactoryGirl)
 
-CypressDev::SmartFactoryWrapper.configure(
+CypressOnRails::SmartFactoryWrapper.configure(
     always_reload: !Rails.configuration.cache_classes,
     factory: factory,
     files: %w(spec/factories.rb ./spec/factories/**/*.rb)